Fix lintors error
