+++ src/Components/layouts/NewBookForm.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { v4 as uuidv4 } from 'uuid';
 import classes from '../../Styles.module.css';
-import { add, addnewBookFunction } from '../../redux/books/books';
+import { addnewBookFunction } from '../../redux/books/books';
 
 const propTypes = {};
