+++ src/Components/pages/Category.js
@@ -1,13 +1,10 @@
 import React from 'react';
-import { useDispatch } from 'react-redux';
-import { add } from '../../redux/books/books';
 
 const propTypes = {};
 
 const defaultProps = {};
 
 const Category = () => {
-  const dispatch = useDispatch();
   const handleTestRedux = () => {
     // dispatch(add());
   };
